Add files via upload
This is the file I re-uploaded after the file with missing VOC data. This file has all VOCs entered and we did not remove compounds with less than X% of samples producing it. I removed all unnecessary data that will not be analyzed in the final analyses and adjusted "sex" for individuals we did not have a molecular sex. The sex of each individual was determined both using molecular techniques and also manual dissections. In doing this, we were also able to confirm there was no contamination during our VOC collections.