Add files via upload
This includes the 3 clones for the growth and development data. For the analyses we only plan on using week 3 measurements (area, perimeter, and circularity) and either total gametophores or days till gametophores. However, you will notice there are Initial - Week 3 measurements for area, perimeter, and circularity and day 7-day 21 measurements for gametophore data. I am leaving this in the data frame in case we want to calculate a growth rate or a developmental rate. Another thing to keep in mind is that although these plants are the same genotypes/individuals as VOCs, the VOC data collection is where we lost the most samples. Thus, there may be samples in this dataset that are not present in the VOC.