Add files via upload
Background only. The backgrounds are much larger than the partial fluxes. It is difficult to plot the partial flux and the background on the same graph. However, we are not comparing the background to the partial flux because that has already been removed from each of the samples anyways. It is more reasonable to only plot the backgrounds for each of the different dates.